chore: op-revm cleanup and few docs
