+++ crates/optimism/src/api/builder.rs
@@ -1,17 +1,21 @@
-use crate::{evm::OpEvm, transaction::OpTxTr, L1BlockInfo, OpSpecId, OpTransaction};
+use crate::{evm::OpEvm, transaction::OpTxTr, L1BlockInfo, OpSpecId};
 use revm::{
-    context::{BlockEnv, Cfg, CfgEnv, JournalOutput, TxEnv},
+    context::{Cfg, JournalOutput},
     context_interface::{Block, JournalTr},
     handler::instructions::EthInstructions,
     interpreter::interpreter::EthInterpreter,
-    Context, Database, Journal,
+    Context, Database,
 };
 
+/// Trait that allows for optimism OpEvm to be built.
 pub trait OpBuilder: Sized {
+    /// Type of the context.
     type Context;
 
+    /// Build the op.
     fn build_op(self) -> OpEvm<Self::Context, (), EthInstructions<EthInterpreter, Self::Context>>;
 
+    /// Build the op with an inspector.
     fn build_op_with_inspector<INSP>(
         self,
         inspector: INSP,
@@ -39,6 +43,3 @@ where
         OpEvm::new(self, inspector)
     }
 }
-
-pub type OpContext<DB> =
-    Context<BlockEnv, OpTransaction<TxEnv>, CfgEnv<OpSpecId>, DB, Journal<DB>, L1BlockInfo>;
